refactor(add): Officially define an op
