+++ crates/cargo-add/src/ops/cargo_add/crate_spec.rs
@@ -64,20 +64,6 @@ impl CrateSpec {
         Ok(id)
     }
 
-    /// Whether the version req is known or not
-    pub fn has_version(&self) -> bool {
-        match self {
-            Self::PkgId {
-                name: _,
-                version_req,
-            } => version_req.is_some(),
-            Self::Path(_path) => {
-                // We'll get it from the manifest
-                true
-            }
-        }
-    }
-
     /// Generate a dependency entry for this crate specifier
     pub fn to_dependency(&self) -> CargoResult<Dependency> {
         let dep = match self {
