+++ crates/cargo-add/src/ops/mod.rs
@@ -1,3 +1,8 @@
 //! Core of cargo commands
 
 pub mod cargo_add;
+
+pub use self::cargo_add::add;
+pub use self::cargo_add::AddOptions;
+pub use self::cargo_add::DepOp;
+pub use self::cargo_add::DepTable;
